Move general_functions
